Chess: refactor board background logic to stand alone class
